+++ services-turf/src/test/java/com/mapbox/turf/TurfTransformationSimplifySpecialCasesTest.java
@@ -10,19 +10,19 @@ import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 
-public class TurfSimplifySpecialCasesTest {
+public class TurfTransformationSimplifySpecialCasesTest {
 
     @Test
     public void emptyCoordinates() {
         List<Point> input = new ArrayList<>();
-        List<Point> output = TurfSimplify.simplify(input, 0.01, true);
+        List<Point> output = TurfTransformation.simplify(input, 0.01, true);
         assertEquals(0, output.size());
     }
 
     @Test
     public void oneCoordinate() {
         List<Point> input = Collections.singletonList(Point.fromLngLat(1.0, 2.0));
-        List<Point> output = TurfSimplify.simplify(input, 0.01, false);
+        List<Point> output = TurfTransformation.simplify(input, 0.01, false);
         assertEquals(input, output);
     }
 
@@ -32,7 +32,7 @@ public class TurfSimplifySpecialCasesTest {
                 Point.fromLngLat(1.0, 2.0),
                 Point.fromLngLat(-56.12, 87.09)
         );
-        List<Point> output = TurfSimplify.simplify(input, 0.01, true);
+        List<Point> output = TurfTransformation.simplify(input, 0.01, true);
         assertEquals(input, output);
     }
 }